requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# Anthropic Claude API
anthropic>=0.18.0

# Scheduling
apscheduler>=3.10.0

# PDF Processing
pdfplumber>=0.10.0

# Environment & Configuration
python-dotenv>=1.0.0

# Async SQLite
aiosqlite>=0.19.0

# Timezone handling
pytz>=2024.1

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
